migrate/oldwiki/requirements.txt
requests
beautifulsoup4
PyYAML
html-to-markdown
